+++ dwds/lib/data/connect_request.g.dart
@@ -81,11 +81,11 @@ class _$ConnectRequest extends ConnectRequest {
       required this.instanceId,
       required this.entrypointPath})
       : super._() {
-    BuiltValueNullFieldError.checkNotNull(appId, 'ConnectRequest', 'appId');
+    BuiltValueNullFieldError.checkNotNull(appId, r'ConnectRequest', 'appId');
     BuiltValueNullFieldError.checkNotNull(
-        instanceId, 'ConnectRequest', 'instanceId');
+        instanceId, r'ConnectRequest', 'instanceId');
     BuiltValueNullFieldError.checkNotNull(
-        entrypointPath, 'ConnectRequest', 'entrypointPath');
+        entrypointPath, r'ConnectRequest', 'entrypointPath');
   }
 
   @override
@@ -113,7 +113,7 @@ class _$ConnectRequest extends ConnectRequest {
 
   @override
   String toString() {
-    return (newBuiltValueToStringHelper('ConnectRequest')
+    return (newBuiltValueToStringHelper(r'ConnectRequest')
           ..add('appId', appId)
           ..add('instanceId', instanceId)
           ..add('entrypointPath', entrypointPath))
@@ -169,14 +169,14 @@ class ConnectRequestBuilder
     final _$result = _$v ??
         new _$ConnectRequest._(
             appId: BuiltValueNullFieldError.checkNotNull(
-                appId, 'ConnectRequest', 'appId'),
+                appId, r'ConnectRequest', 'appId'),
             instanceId: BuiltValueNullFieldError.checkNotNull(
-                instanceId, 'ConnectRequest', 'instanceId'),
+                instanceId, r'ConnectRequest', 'instanceId'),
             entrypointPath: BuiltValueNullFieldError.checkNotNull(
-                entrypointPath, 'ConnectRequest', 'entrypointPath'));
+                entrypointPath, r'ConnectRequest', 'entrypointPath'));
     replace(_$result);
     return _$result;
   }
 }
 
-// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,no_leading_underscores_for_local_identifiers,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
+// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,no_leading_underscores_for_local_identifiers,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new,unnecessary_lambdas