+++ dwds/lib/data/register_event.g.dart
@@ -69,9 +69,9 @@ class _$RegisterEvent extends RegisterEvent {
   _$RegisterEvent._({required this.eventData, required this.timestamp})
       : super._() {
     BuiltValueNullFieldError.checkNotNull(
-        eventData, 'RegisterEvent', 'eventData');
+        eventData, r'RegisterEvent', 'eventData');
     BuiltValueNullFieldError.checkNotNull(
-        timestamp, 'RegisterEvent', 'timestamp');
+        timestamp, r'RegisterEvent', 'timestamp');
   }
 
   @override
@@ -96,7 +96,7 @@ class _$RegisterEvent extends RegisterEvent {
 
   @override
   String toString() {
-    return (newBuiltValueToStringHelper('RegisterEvent')
+    return (newBuiltValueToStringHelper(r'RegisterEvent')
           ..add('eventData', eventData)
           ..add('timestamp', timestamp))
         .toString();
@@ -145,12 +145,12 @@ class RegisterEventBuilder
     final _$result = _$v ??
         new _$RegisterEvent._(
             eventData: BuiltValueNullFieldError.checkNotNull(
-                eventData, 'RegisterEvent', 'eventData'),
+                eventData, r'RegisterEvent', 'eventData'),
             timestamp: BuiltValueNullFieldError.checkNotNull(
-                timestamp, 'RegisterEvent', 'timestamp'));
+                timestamp, r'RegisterEvent', 'timestamp'));
     replace(_$result);
     return _$result;
   }
 }
 
-// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,no_leading_underscores_for_local_identifiers,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new
+// ignore_for_file: always_put_control_body_on_new_line,always_specify_types,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,deprecated_member_use_from_same_package,lines_longer_than_80_chars,no_leading_underscores_for_local_identifiers,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new,unnecessary_lambdas